update bill and Gas

+++ IstaPay/Bill.java
@@ -1,5 +1,6 @@
 package IstaPay;
 
 public interface  Bill extends Transfer {
-    public float calcAmount(float usage);
+    //public float calcAmount(float usage);
+    public double calcAmount(double usage);
 }
